ml/requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
numpy==2.1.2
psycopg2-binary==2.9.9

